+++ Changelog.md
@@ -1,6 +1,7 @@
 # Changelog - Releases
 
 ## Unreleased
+- use wp_safe_redirect for redirects [#115](https://github.com/nicumicle/simple-jwt-login/issues/115)
 
 ## 3.5.7 (22 Dec 2024)
 - Update WordPress 6.7 Compatibility